franchise-enrichment/requirements.txt
pandas
openpyxl
openai
python-dotenv
playwright
pydantic
stagehand
google-cloud-storage